Server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "fastapi-limiter>=0.1.6",
    "fastapi[all]>=0.115.5",
    "logtail-python>=0.3.1",
    "loguru>=0.7.2",
    "mongoengine>=0.29.1",
    "motor>=3.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "phonenumbers>=8.13.50",
    "pydantic-extra-types>=2.10.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.17",
    "requests>=2.32.3",
    "uvicorn>=0.32.1",
]

[dependency-groups]
dev = [
    "mongomock>=4.3.0",
    "pylint-mongoengine>=0.6.0",
    "pylint-pytest>=1.1.7",
    "pytest>=8.3.3",
]
